Refactor HTMLElement: moved the add method
